MGRV-78 - behavior for main drop down menus
