+++ styleguide/source/_patterns/03-organisms/header.twig
@@ -8,12 +8,12 @@
     <section class="ma__header__utility-nav">
         <nav class="ma__utility-nav js-util-nav">
             <ul class="ma__utility-nav__items">
-                <li class="ma__utility-nav__item js-toggle-util-nav">
+                <li class="ma__utility-nav__item js-util-nav-toggle">
                     <a class="ma__utility-nav__link" href="#" title="State Services">State Services</a>
                 </li>
-                <li class="ma__utility-nav__item js-toggle-util-nav"><a class="ma__utility-nav__link" href="#" title="Alerts">Alerts</a></li>
-                <li class="ma__utility-nav__item js-toggle-util-nav"><a class="ma__utility-nav__link" href="#" title="Site Settings">Site Settings</a></li>
-                <li class="ma__utility-nav__item js-toggle-util-nav"><a class="ma__utility-nav__link" href="#" title="login to ...">Login to ...</a></li>
+                <li class="ma__utility-nav__item js-util-nav-toggle"><a class="ma__utility-nav__link" href="#" title="Alerts">Alerts</a></li>
+                <li class="ma__utility-nav__item js-util-nav-toggle"><a class="ma__utility-nav__link" href="#" title="Site Settings">Site Settings</a></li>
+                <li class="ma__utility-nav__item js-util-nav-toggle"><a class="ma__utility-nav__link" href="#" title="login to ...">Login to ...</a></li>
             </ul>
             <ul class="ma__utility-nav__content-boxes">
                 <li class="ma__utility-nav__content-box js-util-nav-content is-closed"></li>
@@ -45,10 +45,11 @@
     </section>
     <section class="ma__header__nav">
         <div>
-        <nav class="ma__main-nav">
-            <ul class="ma__main-nav__items">
-                <li class="ma__main-nav__item"><a class="ma__main-nav__link" href="#" title="Living">Living</a>
-                    <ul class="ma__main-nav__sub-items">
+        <nav class="ma__main-nav ">
+            <ul class="ma__main-nav__items js-main-nav">
+                <li class="ma__main-nav__item js-main-nav-toggle">
+                    <a class="ma__main-nav__top-link" href="#" title="Living">Living</a>
+                    <ul class="ma__main-nav__sub-items js-main-nav-content is-closed">
                         <li class="ma__main-nav__subitem"><a href="" class="ma__main-nav__link">Lorem ipsum.</a></li>
                         <li class="ma__main-nav__subitem"><a href="" class="ma__main-nav__link">Culpa, beatae?</a></li>
                         <li class="ma__main-nav__subitem"><a href="" class="ma__main-nav__link">Facere, aspernatur.</a></li>
@@ -56,8 +57,9 @@
                         <li class="ma__main-nav__subitem"><a href="" class="ma__main-nav__link">Excepturi, perferendis.</a></li>
                     </ul>
                 </li>
-                <li class="ma__main-nav__item"><a class="ma__main-nav__link" href="#" title="Working">Working</a>
-                    <ul class="ma__main-nav__sub-items">
+                <li class="ma__main-nav__item js-main-nav-toggle">
+                    <a class="ma__main-nav__top-link" href="#" title="Working">Working</a>
+                    <ul class="ma__main-nav__sub-items js-main-nav-content is-closed">
                         <li class="ma__main-nav__subitem"><a href="" class="ma__main-nav__link">Lorem ipsum.</a></li>
                         <li class="ma__main-nav__subitem"><a href="" class="ma__main-nav__link">Dolore, ipsum?</a></li>
                         <li class="ma__main-nav__subitem"><a href="" class="ma__main-nav__link">Voluptatibus, rerum?</a></li>
@@ -67,23 +69,26 @@
                         <li class="ma__main-nav__subitem"><a href="" class="ma__main-nav__link">Numquam, doloremque.</a></li>
                     </ul>
                 </li>
-                <li class="ma__main-nav__item"><a class="ma__main-nav__link" href="#" title="Learning">Learning</a>
-                    <ul class="ma__main-nav__sub-items">
+                <li class="ma__main-nav__item js-main-nav-toggle">
+                    <a class="ma__main-nav__top-link" href="#" title="Learning">Learning</a>
+                    <ul class="ma__main-nav__sub-items js-main-nav-content is-closed">
                         <li class="ma__main-nav__subitem"><a href="" class="ma__main-nav__link">Lorem ipsum.</a></li>
                         <li class="ma__main-nav__subitem"><a href="" class="ma__main-nav__link">Tempore, quos!</a></li>
                         <li class="ma__main-nav__subitem"><a href="" class="ma__main-nav__link">Hic, numquam.</a></li>
                     </ul>
                 </li>
-                <li class="ma__main-nav__item"><a class="ma__main-nav__link" href="#" title="Visiting">Visiting &amp; Exploring</a>
-                    <ul class="ma__main-nav__sub-items">
+                <li class="ma__main-nav__item js-main-nav-toggle">
+                    <a class="ma__main-nav__top-link" href="#" title="Visiting">Visiting &amp; Exploring</a>
+                    <ul class="ma__main-nav__sub-items js-main-nav-content is-closed">
                         <li class="ma__main-nav__subitem"><a href="" class="ma__main-nav__link">Lorem ipsum.</a></li>
                         <li class="ma__main-nav__subitem"><a href="" class="ma__main-nav__link">Commodi, voluptatem.</a></li>
                         <li class="ma__main-nav__subitem"><a href="" class="ma__main-nav__link">Officia, pariatur.</a></li>
                         <li class="ma__main-nav__subitem"><a href="" class="ma__main-nav__link">Alias, possimus.</a></li>
                     </ul>
                 </li>
-                <li class="ma__main-nav__item"><a class="ma__main-nav__link" href="#" title="Your Government">Your Government</a>
-                    <ul class="ma__main-nav__sub-items">
+                <li class="ma__main-nav__item js-main-nav-toggle">
+                    <a class="ma__main-nav__top-link" href="#" title="Your Government">Your Government</a>
+                    <ul class="ma__main-nav__sub-items js-main-nav-content is-closed">
                         <li class="ma__main-nav__subitem"><a href="" class="ma__main-nav__link">Lorem ipsum.</a></li>
                         <li class="ma__main-nav__subitem"><a href="" class="ma__main-nav__link">Saepe, repellendus.</a></li>
                         <li class="ma__main-nav__subitem"><a href="" class="ma__main-nav__link">Cumque, quae!</a></li>
